chore: Update Angular and npm dependencies
This commit updates the Angular and npm dependencies in the project. It includes updating the `@angular/cli` version to `^18.0.1` and the `@angular/compiler-cli` version to `^18.0.0`. Additionally, it adds new dependencies like `autoprefixer` version `^10.4.19`, `postcss` version `^8.4.38`, and `tailwindcss` version `^3.4.3`.

The changes in this commit ensure that the project is up to date with the latest stable versions of the dependencies, which can improve performance, security, and compatibility.

Please review and merge this commit to keep the project dependencies up to date.