refactor: Update user component to use ngFor directive for rendering users
This commit updates the user component to use the ngFor directive for rendering the list of users. Instead of manually selecting a random user from the list of dummy users, the component now iterates over the users array and renders each user using the app-user component. This change improves the maintainability and scalability of the code by leveraging Angular's built-in directive for handling list rendering.